+++ WalletWasabi.Backend/Global.cs
@@ -10,9 +10,9 @@ using WalletWasabi.BitcoinCore.Mempool;
 using WalletWasabi.BitcoinCore.Rpc;
 using WalletWasabi.Blockchain.BlockFilters;
 using WalletWasabi.Blockchain.Blocks;
+using WalletWasabi.Discoverability;
 using WalletWasabi.Helpers;
 using WalletWasabi.Logging;
-using WalletWasabi.Nostr;
 using WalletWasabi.Services;
 using WalletWasabi.WabiSabi;
 using WalletWasabi.WabiSabi.Backend;
