+++ WalletWasabi/Discoverability/NostrCoordinatorConfiguration.cs
@@ -1,5 +1,5 @@
 using NBitcoin;
 
-namespace WalletWasabi.Nostr;
+namespace WalletWasabi.Discoverability;
 
 public record NostrCoordinatorConfiguration(string Description, Uri Uri, Network Network);
